PKTT: add Pktt utilities module
This module has classes which augment functionality
of basic data types like displaying integers as their
hex equivalent. It also includes an Enum base class
which displays the integer as its string representation
given by a mapping dictionary. There is also a class
to be used as a base class for an RPC payload object.
This module also includes some module variables to
change how certain objects are displayed.