TOOLS: add tool process_xdr.py
It converts XDR definition file into python code.
A couple of files are created: xdrfile_const.py where
all constant definitions and enum dictionaries are
stored and xdrfile.py where the python code
corresponding to all structures and discriminated
unions are stored.

In addition to processing the XDR definitions,
it processes different tags to change or expand
the behavior of the python object being created.
These tags are given as comments in the XDR
definition file