PKTT: use well defined function prototypes
The use of well defined prototypes instead of having
an open interface increases performance due to the
fact that method arguments are not processed twice
looking at the positional array and the named
arguments dictionary.
Also, the method argument "args" was renamed to
"uargs" to avoid using the args word which has a
special meaning in the python debugger.