PROCESS_XDR: add processing for module description
Processing the XDR file, a comment marker or tag
is added for processing a description for both the
decoding and constants modules. The marker is called
 __DESCRIPTION__ for the decoding module and within
the comment a separate marker __CONST__ is used for
the description for the constants module.
Also, a default description is used when the marker
does not exist