TOOLS: add packet trace decoder tool
Decode and display all packets in the packet trace
file(s) given. The match option gives the ability
to search for specific packets within the packet
trace file. Other options allow displaying of their
corresponding call or reply when only one or the
other is matched. Only a range of packets can be
displayed if the start and/or end options are used.