TOOLS: add tool to display all packets for a specific file
Display all NFS packets for the specified path. It takes a
relative path, where it searches for each of the directory
entries given in the path until it gets the file handle for
the directory where the file is located. Once the directory
file handle is found, a LOOKUP or OPEN/CREATE is searched
for the given file name. If the file lookup or creation is
found, all file handles and state ids associated with that
file are searched and all packets found, including their
respective replies are displayed.