INTEROP: initial commit for interoperability tests
Basic interoperability tests verify that a file written
with different versions of NFS is written correctly.
The contents of the file are verified by reading the
file back using one of the NFS versions.

The tests append different data from different versions
of NFS one at a time then reads the contents of the file
to verify it was written correctly. This is done twice
for each test.