TestUtil: add option to create sparse files
The ftype option is added to create_file method to create
a sparse file. A sparse file can be created using either
by offset in which a hole is created when a file is not
written to the hole segment or by using deallocate.
There is also another option to initialize the hole to zeros