NFSUtil: add method to get the correct I/O stateid
Add method get_stateid to search for the correct
stateid to use in I/O operations. It searches for
and OPEN for the given file name and get the open
and delegation stateids then it looks for a LOCK
and decides which is the correct stateid to use
on I/O operations.

Add method to display a stateid in CRC16 format.

Add method get_freebytes to get the number of bytes
available in the given directory. It takes into
account the user id because the root user is allowed
to use all disk space while non-root users are
allowed a little bit less.