ALLOC: initial commit for space reservation tests
Verify correct functionality of space reservations so
applications are able to reserve or unreserve space for
a file. The system call fallocate is used to manipulate
the allocated disk space for a file, either to preallocate
or deallocate it. For filesystems which support the
fallocate system call, preallocation is done quickly
by allocating blocks and marking them as uninitialized,
requiring no I/O to the data blocks. This is much faster
than creating a file and filling it with zeros.