SPARSE: initial commit for sparse file tests
Verify correct functionality of sparse files.
The new NFSv4.2 operation SEEK is used to search
for the next hole or data segment in a file.

Basic tests verify the SEEK operation returns the
correct offset of the next hole or data with respect
to the starting offset given to the seek system call.
Verify the SEEK operation is sent to the server with
the correct stateid as a READ call. All files have a
virtual hole at the end of the file so when searching
for the next hole, even if the file does not have a
hole, it returns the size of the file.