PKTT: use up all RPC load bytes when unable to decode
When a packet trace has a lot of frames with multiple
NFS packets and if one of them is not decoded correctly
it could lead to an unsynchronized TCP stream causing
a lot of NFS packets being left undecoded. When an RPC
load is not abled to be decoded at least use up all
the bytes so the TCP stream be in sync.