NFSUtil: verify pNFS is supported for mounted path
Verify pNFS is suported for each directory segment
in mounted path including datadir. It looks for the
LOOKUP to get the file handle and then the GETATTR
for the file handle to get the supported attributes.
It verifies FATTR4_FS_LAYOUT_TYPES is supported for
the path segment. Then it finds the GETATTR asking
for FATTR4_FS_LAYOUT_TYPES(bit 62) to verify
LAYOUT4_NFSV4_1_FILES is returned in fs_layout_types.