REXEC: change remote server to an object
The remote server is changed to an object which removes
the need for a bare-bones server by executing a minimal
python script to run the remote server source code which
is sent via the standard input instead.
Add sudo option to run the remote server as root
so it can be used to run as a different user by
using seteuid.