ALLOC: check for NOSPC errors on OPEN
When writing to another file when there is no more space
available check for errors on OPEN when the WRITE has
not been found.
Also, check if deleg_stateid attribute is available before
using it -- this could happen if there is an error on the
previous test where this attribute should be defined.