PKTT: add buffered matching to increase performance
Add method set_pktlist to set a list of packets in
which the match method will only use this list instead
of getting the next packet from the packet trace file.
This could be used when there is a lot of matching
going back and forth but only on a particular set of
packets.
Also add method get_index the return the current packet
index either from the new buffered packet list or the
actual index from the packet trace file.