NFSUtil: add method set_pktlist to enable buffered matching
This method creates a list of packets to be used for
buffered matching in which the match method will only
use this list instead of getting the next packet from
the packet trace file. The default is to get all packets
unless any of the arguments is given.
Also, fix method verify_io to make sure it works with
buffered matching.