PKTT: add method add_layer to Pkt class
Use add_layer to add each layer object to the Pkt object
and also append object name to a list for display purposes.
This is exactly the same as using set_attr() but a lot
faster since set_attr is very generic with more options
and more error checking which is not needed here.