IB: add reassembly for the RPC-over-RDMA reply chunk
In general, the reassembled message consists of one or multiple
chunks and each chunk in turn could be composed of multiple
segments. Also, each segment could be composed of multiple
sub-segments and each sub-segment could be composed of multiple
fragments.

The protocol only defines segments but if the segment length
is large, it is split into multiple sub-segments in which each
sub-segment is specified by RDMA_WRITE_First or RDMA_READ_Request
packets. The handle is the same for each of these packets but
with a shorter DMA length.

Thus in order to reassemble all fragments for a single message,
a list of segments is created where each segment is identified
by its handle or RKey and the message is reassembled according
to the chuck lists specified by the RPC-over-RDMA layer.

The RPCoRDMA reply is used for a large RPC reply which does
not fit into a single SEND operation and does not have a single
DDP eligible item or large opaque, e.g., NFS READDIR.
The RPCoRDMA reply is just a single write chunk if it exists.
Each segment is a write or plain segment defined by a handle
and its DMA length.

The RPCoRDMA reply is processed on the call message to set up
the reply chunk and its respective segments expected by the
reply. The whole RPC reply is transferred via RDMA writes.
The RPC reply packet has no data, it is sent with RPCoRDMA
proc of RDMA_NOMSG and all the fragments are then reassembled
and the whole RPC reply is dissected.