PROCESS_XDR: add FWRAP modifier
FWRAP: attr=fname[,attr1=fname1[...]]
Add function wrapper to the given attribute definition
In order to add wrapper correctly, the decoding
definition should not be written for each different
type but save the definition and then write it at
the end. This way also the STRHEX tag will be added
to any definition including on top of the wrapper.