NFSUtil: change return value in find_layoutget method
Do not return the layout information anymore instead
save it in the object attribute "layout". Also, remove
duplicate code when initializing the layout information
dictionary, set first the general info and then the
information dependent on the type of layout.
Change the return value in verify_layoutget method as
well.