DELEG: verify I/O packets for basic and recall tests
Add methods to verify I/O packets: check for the correct
stateid on the call, the status value on the reply, and
the correct server they are sent to if pNFS is enabled.

In some recall operations like REMOVE, the server may return
an error on I/O sent after the delegation has been returned.
Make sure the correct error is returned if an error is
expected.