DELEG: add pending open tests
Add delegation tests to trigger CLAIM_DELEGATE_CUR on CB_RECALL.
Recall delegation by writing from a second client while having
a second pending open on the first client. When the delegation
is recalled the client should send an OPEN before returning
the delegation. The OPEN is only sent if the pending open mode
is different than the OPEN having the delegation.
For the case where the client is holding a read delegation and
the pending open is for writing the delegation is returned by
the client when the second open is done so there is no
delegation recall.