NFSUtil: make methods work correctly for NFSv3
Most methods now do a no-op and return immediately for NFSv3.
For other methods like find_open, search for LOOKUP or CREATE
instead of OPEN to get the file handle for NFSv3. This allows
for just minor changes to scripts in order to support NFSv3.
In other methods like verify_commit, the only change is to
look for the correct procedure number using the NFSPROC3_*
instead of the OP_* constants.