PKTT: add method to close trace file and do some cleanup
Gracefully close the packet trace file and cleanup attributes
having the most memory allocated. The Python garbage collection
is not calling the Pktt destructor because the reference count
is not 0 when the Pktt object is dereference, thus when opening
a large number of packet traces the memory is not being freed
even though the objects are no longer referenced. This method
could be called explicitly to force the garbage collector to
free some memory.