NFSTEST_PKT: add option to process trace files serially
When option --serial is given the packet traces are processed
one after the other in the order in which they are given. The
default is to open all files first and then display the packets
ordered according to their timestamps. Call the close method
after each file to force the garbage collector to free all the
memory associated with each Pktt object. This way the script
can handle thousands of files without running out of memory.