TestUtil: add wrapper to trace_open to log NFS errors
Add option --nfserrors to log NFS errors found on the packet
traces. The new trace_open method defined here first calls
the original trace_open method and then searches the packet
trace for any NFS errors and logs them as failures.
A new method set_nfserr_list is defined to temporarily set
the expected NFS errors for the next trace file. There is
a list of expected errors that could be set for NFSv3, NFSv4,
NLM4 and MOUNT3. The default is to ignore NOENT errors for
both NFSv3 and NFSv4.