DELEG: lock file in the second process
For tests with a pending open, lock the file as well in the
second process so two locks will be recovered when the
delegation is recalled. Search for both locks in the packet
trace and distinguished each lock by its lock range.

Change the length of the lock to a finite amount instead of
locking up to the end of the file so the second lock can be
granted. Add another test option --lock-poffset for the
offset of the lock in the second process.