IB: change RDMAinfo to an instance attribute
RDMAinfo as a class attribute will cause problems when
having multiple Pktt instances since changing this
attribute in one instance all other instances will see
the change. The purpose of RDMAinfo is so all IB packets
can have access to the RDMA segments but within the
same Pktt instance not across different Pktt instances.
Changing the RDMAinfo object from a class attribute
of IB to an instance attribute of Pktt this issue is
avoided.