TestUtil: add user option client-ipaddr
This option is used for debugging purposes specifically
when debugging test scripts coming from a different
client where the IP address do not match with the
machine under test. Use in conjunction with the option
basename which does not execute the test but just looks
at the packet traces from a previous run.