TestUtil: add trace marker functionality
Add method insert_trace_marker to add a marker in the
packet trace by inserting a LOOKUP for a non-existent
file. Each time this method is called a unique trace
marker is generated. If the marker name is given the
caller must make sure this is a unique name in order
to find the correct index for this marker. This could
also be used to add any arbitrary information to the
packet trace.

Add method get_marker_index to find the packet index
for a trace marker. If the trace marker id is not
given the current marker id is used.

Add a wrapper to trace_start in order to reset the
trace marker state.