PNFS: do not read the whole file in rw tests
In the rw_read and rw_write tests, read just the amount
of bytes specified in the --rsize option instead of
reading the whole file.

Also, add a trace marker after the first part of the
test to make sure the second part of the test does not
match any incorrect packets.