LOCK: add ProcInfo class to bundle Rexec and Host
The ProcInfo class has both the Rexec object and the Host
object for the given process. This way each process can
have its own Host object which means it could now have
different mount options including a different server and
export directory.

Also, this class has attributes to have the locking info
for the given process, in which this include the lock
range (offset and length), the file descriptor to make
it easier to close the file for each process, etc.

A ProcInfo object is a direct replacement for the Rexec
object since attributes and methods from Rexec can be
called directly using the ProcInfo object.