LOCK: add support for multiple locks to wait_for_lock
Use a list of ProcInfo objects with info for each blocked
lock on a different process or client as an argument to
this wait_for_lock instead of a single execobj.

Also, use a different message for the current lock and
a list of messages for each lock given in the ProcInfo
list.