LOCK: add support for multiple locks to do_basic_lock
Use a list of ProcInfo objects with info for each extra
lock on a different process or client as an argument to
do_basic_lock instead of a single lock definition
specified by execobj, offset2 and length2.

Use get_info_list to get a list of ProcInfo objects
given a list of extra lock ranges.