NFSUtil: add nfs_version option to find_open
Add option to set the NFS version to search for the OPEN
if NFSv4.x or a LOOKUP for the case of NFSv3. The default
value for this option is the NFS version of the mounted
share. This is used for tests having traffic with multiple
versions of NFS especially NFSv4.x and NFSv3.